cleanup lcavity tests into one directory
